Player in game
The player can now move In the game and can open doors.
we can move with WASD and turn righ with E and left with Q.
notice that this is not in continues motion yet, that is still in the works but for now we have the ability to move and in the furture we want to add this code to the robot that we have designed.